Refactoring (splitting the "create edit item" into components)
