+++ src/templates/create-trip-events-list.js
