+++ src/templates/event-item/create-trip-event-item-edit/create-trip-event-details-wrapper.js
@@ -1,0 +1,5 @@
+export const createTripEventDetailsWrapper = () => (
+  `
+  <section class="event__details"></section>
+  `
+);
