+++ src/templates/event-item/create-trip-event-item.js
