+++ src/templates/header/create-trip-info-header.js
