+++ src/templates/header/create-trip-tabs-header.js
